+++ src/index.js
@@ -2,9 +2,11 @@ import { Router } from './Router.js';
 import { createServerHistory } from './createServerHistory.js';
 import { createBrowserHistory } from './createBrowserHistory.js';
 import { createTestHistory } from './createTestHistory.js';
+import { RouteCollection } from './RouteCollection.js';
 
 export {
   Router,
+  RouteCollection,
   createServerHistory,
   createBrowserHistory,
   createTestHistory,